+++ entrypoint.sh
@@ -5,7 +5,7 @@ set -o errexit
 # pmm-managed-init validates environment variables.
 pmm-managed-init
 
-/usr/share/percona-dashboards/import-dbass-dashboard.py&
+/usr/share/percona-dashboards/import-dashboards.py -d&
 
 # Start supervisor in foreground
 exec supervisord -n -c /etc/supervisord.conf
