+++ import-dashboards.py
@@ -23,7 +23,7 @@ import zipfile
 
 import requests
 
-GRAFANA_DB_DIR             = sys.argv[1] if len(sys.argv) > 1 else '/srv/grafana'
+GRAFANA_DB_DIR             = sys.argv[1] if (len(sys.argv) > 1 and sys.argv[1] != '-d') else '/srv/grafana'
 GRAFANA_IMG_DIR            = '/usr/share/grafana/public/img/'
 GRAFANA_PLUGINS_DIR        = '/var/lib/grafana/plugins/'
 GRAFANA_SOURCE_PLUGINS_DIR = '/usr/share/percona-dashboards/panels/'
@@ -36,6 +36,7 @@ LOGO_FILE                  = '/usr/share/pmm-server/landing-page/img/pmm-logo.sv
 SET_OF_TAGS                = {'Query Analytics': 0, 'OS': 0, 'MySQL': 0, 'MongoDB': 0, 'PostgreSQL': 0, 'Insight': 0, 'PMM': 0}
 YEAR                       = str(datetime.date.today())[:4]
 DBAAS                      = os.getenv('PERCONA_TEST_DBAAS') is not None
+GRAFANA_PROCESS            = 'grafana-server'
 
 CONTENT                    = '''<center>
 <p>MySQL and InnoDB are trademarks of Oracle Corp. Proudly running Percona Server. Copyright (c) 2006-'''+YEAR+''' Percona LLC.</p>
@@ -151,7 +152,7 @@ def add_api_key(name, db_key):
     con.close()
 
 
-def delete_api_key(db_key, upgrade):
+def delete_api_key(db_key):
     con = sqlite3.connect(GRAFANA_DB_DIR + '/grafana.db', isolation_level="EXCLUSIVE")
     cur = con.cursor()
 
@@ -530,21 +531,68 @@ def set_home_dashboard(api_key):
     #             "SELECT 1, 1, 0, 0, id, '', '', datetime('now'), datetime('now') from dashboard WHERE slug='home'")
 
 
-def remove_dbaas_dashboard(api_key):
-    if DBAAS:
-        print ' * DBaaS is enabled'
+def dbaas_dashboard(api_key):
+    if not DBAAS:
+        print ' * DBaaS is disabled'
+        r = requests.get('%s/api/dashboards/uid/pmm-dbaas' % (HOST,), headers=grafana_headers(api_key))
+        if r.status_code == httplib.OK:
+            print '   * Removing DBaaS dashboard'
+            r = requests.delete('%s/api/dashboards/uid/pmm-dbaas' % (HOST,), headers=grafana_headers(api_key))
+            if r.status_code != httplib.OK:
+                print r.status_code, r.content
         return
 
-    print ' * DBaaS is disabled'
+    print ' * DBaaS is enabled'
     r = requests.get('%s/api/dashboards/uid/pmm-dbaas' % (HOST,), headers=grafana_headers(api_key))
-    if r.status_code == httplib.OK:
-        print '   * Removing DBaaS dashboard'
-        r = requests.delete('%s/api/dashboards/uid/pmm-dbaas' % (HOST,), headers=grafana_headers(api_key))
+    if r.status_code != httplib.OK:
+        print '   * Adding DBaaS dashboard'
+        with open('/usr/share/percona-dashboards/pmm-app/dist/dashboards/pmm-dbaas.json', 'r') as dashboard_file:
+            dashboard = json.loads(dashboard_file.read())
+        data = json.dumps({'dashboard': dashboard})
+        r = requests.post('%s/api/dashboards/import' % (HOST,), data=data, headers=grafana_headers(api_key))
         if r.status_code != httplib.OK:
             print r.status_code, r.content
+    else:
+       print '   * DBaaS dashboard has already been added'
+
+
+def check_active_process(processname):
+    ps = os.popen("ps -Af").read()
+    proccount = ps.count(processname)
+
+    if proccount > 0:
+        print'%s service is running' % (processname,)
+        return 1
 
 
 def main():
+    # Add/Remove DBaaS dashboard PMM-7085
+    if (len(sys.argv) > 1 and sys.argv[1] == '-d'):
+        # wait till grafana will be run
+        while not check_active_process(GRAFANA_PROCESS):
+            time.sleep(1)
+            print'no %s service has run yet' % (GRAFANA_PROCESS,)
+        # modify database when Grafana is stopped to avoid a data race
+        stop_grafana()
+        name, api_key, db_key = get_api_key()
+
+        try:
+            add_api_key(name, db_key)
+        finally:
+            start_grafana()
+
+        wait_for_grafana_start()
+
+        dbaas_dashboard(api_key)
+
+        # modify database when Grafana is stopped to avoid a data race
+        stop_grafana()
+        try:
+            delete_api_key(db_key)
+        finally:
+            start_grafana()
+        return
+
     print "Grafana database directory: %s" % (GRAFANA_DB_DIR,)
     upgrade = check_dashboards_version()
 
@@ -567,7 +615,7 @@ def main():
     get_folders(api_key)
     import_apps(api_key)
     move_into_folders()
-    remove_dbaas_dashboard(api_key)
+    dbaas_dashboard(api_key)
 
     # restart Grafana to load app and set home dashboard below
     stop_grafana()
@@ -580,7 +628,7 @@ def main():
     # modify database when Grafana is stopped to avoid a data race
     stop_grafana()
     try:
-        delete_api_key(db_key, upgrade)
+        delete_api_key(db_key)
     finally:
         start_grafana()
